// src/app/models/cityDetails.ts
import { Injectable } from '@angular/core';

@Injectable({
    providedIn: 'root'
})

export class CityDetailsLondon {
    Id: number;
    CityName: string;
    Temperature: string;
    SunriseTime: string;
    SunSetTime: string;
    SeaLevel: number;
}
@Injectable({
    providedIn: 'root'
})

export class CityDetailsManchester {
    Id: number;
    CityName: string;
    Temperature: string;
    SunriseTime: string;
    SunSetTime: string;
    SeaLevel: number;
}
@Injectable({
    providedIn: 'root'
})

export class CityDetailsCambridge {
    Id: number;
    CityName: string;
    Temperature: string;
    SunriseTime: string;
    SunSetTime: string;
    SeaLevel: number;
}
@Injectable({
    providedIn: 'root'
})

export class CityDetailsOxford {
    Id: number;
    CityName: string;
    Temperature: string;
    SunriseTime: string;
    SunSetTime: string;
    SeaLevel: number;
}
@Injectable({
    providedIn: 'root'
})

export class CityDetailsBristol {
    Id: number;
    CityName: string;
    Temperature: string;
    SunriseTime: string;
    SunSetTime: string;
    SeaLevel: number;
}